Update package.json
It's really cool to create containers, but this makes it difficult to start rapid development, which is the proposal for packages of this type. Even if the change is not accepted, I think it's interesting to leave a note for possible new beginners in the future: changes in script calls in package.json of the most current /javascript version of React-Router.